Add AnnotateRoutes::HeaderGenerator.app_routes_map
